requirements.txt
asgiref==3.9.1
astropy==7.1.0
astropy-iers-data==0.2025.10.6.0.35.25
bcrypt==4.3.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.8.3
cffi==1.17.1
cftime==1.6.4.post1
charset-normalizer==3.4.3
contourpy==1.3.3
crispy-bootstrap4==2025.6
cryptography==45.0.5
cycler==0.12.1
digital_rf==2.6.12
Django==5.2.4
django-crispy-forms==2.4
django-filter==25.1
django-tables2==2.7.5
djangorestframework==3.16.0
fonttools==4.60.0
h5py==3.14.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
maidenhead==1.8.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mod_wsgi==5.0.2
mysqlclient==2.2.7
narwhals==2.6.0
netCDF4==1.7.2
numpy==2.3.1
packaging==25.0
pandas==2.3.3
paramiko==3.5.1
pillow==11.3.0
plotly==6.3.0
pvlib==0.13.1
pycparser==2.22
pyerfa==2.0.1.5
PyMySQL==1.1.2
PyNaCl==1.5.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.2
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
xarray==2025.9.1
